[refactor] mv python scripts into `pytorch/`
