+++ pytorch/cnnrnnln.py
@@ -1,5 +1,5 @@
 """
-Profile SARNN inference
+Profile CNNRNNLN inference
 Inference with pretrained weight
 No write result only inference
 Download pretrained before run this script
@@ -8,13 +8,11 @@ import os
 import torch
 import time
 import json
-import sys
 import numpy as np
 import argparse
 from memory_profiler import profile
 
-sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
-from models.sarnn import SARNN
+from models.cnnrnnln import CNNRNNLN
 
 
 argparser = argparse.ArgumentParser()
@@ -38,8 +36,8 @@ else:
     device = torch.device(args.device)
 print("device:{}".format(device))
 
-weight_file = "./downloads/airec/pretrained/SARNN/model.pth"
-params_file = "./downloads/airec/pretrained/SARNN/args.json"
+weight_file = "./downloads/airec/pretrained/CNNRNNLN/model.pth"
+params_file = "./downloads/airec/pretrained/CNNRNNLN/args.json"
 data_dir = "./downloads/airec/grasp_bottle"
 test_data_dir = "./downloads/airec/grasp_bottle/test"
 test_data_index = 0
@@ -59,13 +57,9 @@ joint_bounds = torch.from_numpy(joint_bounds).to(device)
 print("data loaded")
 
 # load model
-model = SARNN(
-    rec_dim=params["rec_dim"],
-    joint_dim=8,
-    k_dim=params["k_dim"],
-    heatmap_size=params["heatmap_size"],
-    temperature=params["temperature"],
-).to(device)
+model = CNNRNNLN(rec_dim=params["rec_dim"], joint_dim=8, feat_dim=params["feat_dim"]).to(
+    device
+)
 
 # load weight
 ckpt = torch.load(weight_file, map_location=device)
@@ -99,7 +93,7 @@ def inference():
         # if PyTorch<1.9 use torch.no_grad() instead
         with torch.inference_mode():
             start_time = time.time()
-            _, _, _, _, state = model(img_t, joint_t, state)
+            _, _, state = model(img_t, joint_t, state)
             end_time = time.time()
         elapsed = end_time - start_time
         time_list.append(elapsed)
