+++ pytorch/export_onnx.py
@@ -1,6 +1,5 @@
 import sys
 import os
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 from models.loader import export_onnx
 
 export_onnx("sarnn", "sarnn.onnx")
